Add main content of Create Profile Component
