+++ package.json
@@ -30,12 +30,12 @@
     "eslint": "^8.2.0",
     "eslint-config-next": "^12.0.1",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-storybook": "^0.5.3",
     "eslint-config-standard": "^16.0.3",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-promise": "^5.1.1",
+    "eslint-plugin-storybook": "^0.5.3",
     "husky": "^7.0.4",
     "lint-staged": "^12.0.2",
     "prettier": "^2.4.1",
